Se generan métodos para el CRUD de pacientes

+++ pacientes/src/main/java/com/minsait/pacientes/controller/PacienteController.java
@@ -4,12 +4,16 @@ import com.minsait.pacientes.model.entity.Direccion;
 import com.minsait.pacientes.model.entity.Paciente;
 import com.minsait.pacientes.repository.PacienteRepository;
 import com.minsait.pacientes.service.PacienteService;
+import jakarta.validation.constraints.NotNull;
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @RestController
@@ -42,13 +46,40 @@ public class PacienteController {
     }
 
     @PostMapping("/insertpaciente")
-    public ResponseEntity<Paciente> savePaciente(@RequestBody Paciente paciente){
-        service.save(paciente);
-        return new ResponseEntity<Paciente>(HttpStatus.CREATED);
+    public ResponseEntity<?> savePaciente(@RequestBody Paciente paciente){
+        ResponseEntity<?> responseEntity= save(paciente);
+        if(responseEntity.getStatusCode().equals(HttpStatus.OK))
+            return ResponseEntity.status(HttpStatus.CREATED).body(responseEntity.getBody());
+        return responseEntity;
     }
-    @DeleteMapping("/deletepaciente")
-    public ResponseEntity<Paciente> savePaciente(@PathVariable Long id){
+
+    @DeleteMapping("/deletepaciente/{id}")
+    public ResponseEntity<?> deletePaciente(@PathVariable Long id){
+        Optional<Paciente> validaDatos= service.findById(id);
+        if (!validaDatos.isPresent())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         service.deleteById(id);
-        return new ResponseEntity<Paciente>(HttpStatus.OK);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PatchMapping("/updatepaciente")
+    public ResponseEntity<?> updatePaciente(@RequestBody Paciente paciente){
+        Optional<Paciente> validaDatos = service.findById(paciente.getId());
+        if (!validaDatos.isPresent())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        ResponseEntity<?> responseEntity= save(paciente);
+        if(responseEntity.getStatusCode().equals(HttpStatus.OK))
+            return ResponseEntity.status(HttpStatus.CREATED).body(responseEntity.getBody());
+        return responseEntity;
+    }
+
+    public ResponseEntity<?> save(@NotNull Paciente paciente){
+        try {
+            return ResponseEntity.ok(service.save(paciente));
+        }catch (Exception e){
+            Map<String, Object> error= new HashMap<>();
+            error.put("Error", e.getMessage());
+            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+        }
     }
 }
